+++ changelog.md
@@ -1,5 +1,9 @@
 # CHANGE LOG
 
+## 1.1.1
+* [ios][fix] RunAll is running inside RunTask producing 2 background tasks /w is also producing bad reporting
+* [ios][fix] RunTask now supports cancellation (only iOS has this functionality)
+
 ## 1.1.0
 * Better internal parameter serializer
 * JobParameters was an unnecessary interface and has been replaced by plain-jane IDictionary /w some beautiful extension methods